Separated browser protocol file back into separate platform-specific files
